refactor(check): relocate GUC duration parser to check.ParseDurationMs
Review feedback on the session-settings parser: there was no existing
duration parser in the library (FormatDurationMs is the formatter; other
checks read already-normalised pg_settings.setting and just ParseInt it),
but it does have a natural home. Move the parser next to its inverse
FormatDurationMs in check/format.go as the exported check.ParseDurationMs,
call it from session-settings, and delete the export_test.go white-box shim
(its only purpose was exposing the unexported helper). Parser unit tests now
live in check/format_test.go — also the first coverage for that file.

No behaviour change.
