+++ check/format.go
@@ -2,6 +2,9 @@ package check
 
 import (
 	"fmt"
+	"math"
+	"strconv"
+	"strings"
 
 	"github.com/jackc/pgx/v5/pgtype"
 )
@@ -54,6 +57,64 @@ func FormatDurationMs(ms float64) string {
 		return fmt.Sprintf("%.1fs", ms/1000)
 	}
 	return fmt.Sprintf("%.0fms", ms)
+}
+
+// timeUnitFactors maps a PostgreSQL GUC time-unit suffix to its millisecond
+// multiplier. PostgreSQL recognises only these lowercase, case-sensitive
+// suffixes; there is no bare "m" time unit ("min" is the only minute spelling).
+var timeUnitFactors = map[string]float64{
+	"us":  1.0 / 1000,
+	"ms":  1,
+	"s":   1000,
+	"min": 60000,
+	"h":   3600000,
+	"d":   86400000,
+}
+
+// ParseDurationMs parses a PostgreSQL GUC time value into integer milliseconds —
+// the inverse of FormatDurationMs. Unlike pg_settings.setting (always normalised
+// to the base unit), raw ALTER ROLE / pg_db_role_setting values are stored
+// verbatim, so they may carry a unit suffix ("2000ms", "1.5s", "2000 ms") or be a
+// bare number interpreted in baseUnit. The sign is preserved so sentinel values
+// like -1 (disabled) survive.
+func ParseDurationMs(value, baseUnit string) (int64, error) {
+	value = strings.TrimSpace(value)
+	if value == "" {
+		return 0, fmt.Errorf("empty value")
+	}
+
+	// Match the longest trailing alphabetic suffix that is a known time unit.
+	numeric, factor := value, durationUnitFactor(baseUnit)
+	for i := len(value) - 1; i >= 0; i-- {
+		if !isASCIILetter(value[i]) {
+			suffix := value[i+1:]
+			if f, ok := timeUnitFactors[suffix]; ok {
+				numeric = strings.TrimSpace(value[:i+1])
+				factor = f
+			}
+			break
+		}
+	}
+
+	n, err := strconv.ParseFloat(numeric, 64)
+	if err != nil {
+		return 0, fmt.Errorf("parsing numeric part %q: %w", numeric, err)
+	}
+
+	return int64(math.Round(n * factor)), nil
+}
+
+// durationUnitFactor returns the ms multiplier for a value's base unit,
+// defaulting to milliseconds when the unit is empty or not a known time unit.
+func durationUnitFactor(unit string) float64 {
+	if f, ok := timeUnitFactors[unit]; ok {
+		return f
+	}
+	return 1
+}
+
+func isASCIILetter(b byte) bool {
+	return (b >= 'a' && b <= 'z') || (b >= 'A' && b <= 'Z')
 }
 
 // FormatDurationSec formats seconds as a human-readable duration (e.g., "2h" or "1d").
