--- checks/sessionsettings/export_test.go
@@ -1,4 +1,0 @@
-package sessionsettings
-
-// ParseDurationMs exposes the unexported duration parser for white-box tests.
-var ParseDurationMs = parseDurationMs
